app/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.2
Flask-Cors==4.0.0
python-dotenv==1.0.0
PyMySQL==1.1.0
pytest==7.4.0
black==23.7.0
flake8==6.1.0 